+++ internal/conformancecheck/main.go
@@ -11,8 +11,8 @@ import (
 )
 
 const (
-	expectedRevision = "1deeb6b87be905edf1ba9b83d8d431def0af702c"
-	expectedVersion  = "0.13"
+	expectedRevision = "a108b314bb3f37622fb082759f726468bbb09163"
+	expectedVersion  = "0.14"
 )
 
 var expectedFixtures = []string{